Removed separate tests for conditions and moved into tests which deploy each component
This change is needed, because "condition" tests have high ration of failures. This is caused
by fact, that each component is created and deleted in each test. In some cases operator can
"stuck" and waits for status of component. But this component is already deleted. Operator then refuses
to create new instance of component. With this change condition tests are run right after
component is available.

+++ functests/06-test-deploy-node-labeller.sh
@@ -29,6 +29,15 @@ for idx in $( seq 1 30); do
 	fi
 	sleep 2s
 done
+
+if [ $RET -eq 1 ] ; then
+	exit 2
+fi
+
+#wait for ssp operator to set proper conditions
+wait_for_condition ${TEST_NS} 5 40 "KubevirtNodeLabellerBundle" "Available" "True"
+RET="$?"
+
 oc delete -n ${TEST_NS} -f "${SCRIPTPATH}/node-labeller-unversioned-cr.yaml" || exit 2
 
 exit $RET
